[third-party/py-spy/0.3.14] Fix chrometrace bug showing dead threads
Summary:
Previously, the trace dumpers would periodically be given per-thread
traces, individually, every period, so if a thread died, it would be difficult
to detect.  As such, the chrometrace impl never emitted end events for
these, and they appeared to never die in the traces.

This fixes by changing py-spy to pass all the traces for all threads to
the trace dumper in one go, which allows it to observe which ones
are no longer there.

Test Plan:
Test profiling a simple test script that runs a thread for 10s, then
runs for another 5s after the thread dies, and verified we no longer
make it look like the thread continued to run.

Before: https://pxl.cl/4f8Dv
After: https://pxl.cl/4f8Dd

Reviewers: bmaurer, kunalb

Reviewed By: bmaurer

Subscribers: mzlee

Differential Revision: https://phabricator.intern.facebook.com/D53142371